tests: reorganize files and functions
